portfolio, support exception service로 변경 #2
