+++ src/main/java/com/tsp/new_tsp_front/api/portfolio/service/impl/FrontPortFolioJpaRepository.java
@@ -4,7 +4,6 @@ import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import com.tsp.new_tsp_front.api.portfolio.domain.FrontPortFolioDTO;
 import com.tsp.new_tsp_front.api.portfolio.domain.FrontPortFolioEntity;
-import com.tsp.new_tsp_front.exception.TspException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
@@ -13,10 +12,9 @@ import java.util.Map;
 
 import static com.tsp.new_tsp_front.api.common.domain.QCommonImageEntity.commonImageEntity;
 import static com.tsp.new_tsp_front.api.portfolio.domain.QFrontPortFolioEntity.frontPortFolioEntity;
+import static com.tsp.new_tsp_front.api.portfolio.service.impl.PortFolioMapper.INSTANCE;
 import static com.tsp.new_tsp_front.common.utils.StringUtil.getInt;
 import static com.tsp.new_tsp_front.common.utils.StringUtil.getString;
-import static com.tsp.new_tsp_front.exception.ApiExceptionType.NOT_FOUND_PORTFOLIO;
-import static com.tsp.new_tsp_front.exception.ApiExceptionType.NOT_FOUND_PORTFOLIO_LIST;
 
 @Repository
 @RequiredArgsConstructor
@@ -47,23 +45,19 @@ public class FrontPortFolioJpaRepository {
 	 * </pre>
 	 *
 	 */
-	public List<FrontPortFolioDTO> getPortFolioList(Map<String, Object> portFolioMap) throws TspException {
-		try {
-			List<FrontPortFolioEntity> portFolioList = queryFactory
-					.selectFrom(frontPortFolioEntity)
-					.where(searchPortFolio(portFolioMap))
-					.orderBy(frontPortFolioEntity.idx.desc())
-					.offset(getInt(portFolioMap.get("jpaStartPage"),0))
-					.limit(getInt(portFolioMap.get("size"),0))
-					.fetch();
+	public List<FrontPortFolioDTO> getPortFolioList(Map<String, Object> portFolioMap) {
+		List<FrontPortFolioEntity> portFolioList = queryFactory
+				.selectFrom(frontPortFolioEntity)
+				.where(searchPortFolio(portFolioMap))
+				.orderBy(frontPortFolioEntity.idx.desc())
+				.offset(getInt(portFolioMap.get("jpaStartPage"),0))
+				.limit(getInt(portFolioMap.get("size"),0))
+				.fetch();
 
-			portFolioList.forEach(list -> portFolioList.get(portFolioList.indexOf(list))
-					.setRnum(getInt(portFolioMap.get("startPage"),1)*(getInt(portFolioMap.get("size"),1))-(2-portFolioList.indexOf(list))));
+		portFolioList.forEach(list -> portFolioList.get(portFolioList.indexOf(list))
+				.setRnum(getInt(portFolioMap.get("startPage"),1)*(getInt(portFolioMap.get("size"),1))-(2-portFolioList.indexOf(list))));
 
-			return PortFolioMapper.INSTANCE.toDtoList(portFolioList);
-		} catch (Exception e) {
-			throw new TspException(NOT_FOUND_PORTFOLIO_LIST, e);
-		}
+		return INSTANCE.toDtoList(portFolioList);
 	}
 
 	/**
@@ -76,20 +70,16 @@ public class FrontPortFolioJpaRepository {
 	 * </pre>
 	 *
 	 */
-	public FrontPortFolioDTO getPortFolioInfo(FrontPortFolioEntity existFrontPortFolioEntity) throws TspException {
-		try {
-			FrontPortFolioEntity getPortFolioInfo = queryFactory
-					.selectFrom(frontPortFolioEntity)
-					.leftJoin(frontPortFolioEntity.commonImageEntityList, commonImageEntity)
-					.fetchJoin()
-					.where(frontPortFolioEntity.idx.eq(existFrontPortFolioEntity.getIdx())
-							.and(frontPortFolioEntity.visible.eq("Y"))
-							.and(commonImageEntity.typeName.eq("portfolio")))
-					.fetchOne();
+	public FrontPortFolioDTO getPortFolioInfo(FrontPortFolioEntity existFrontPortFolioEntity) {
+		FrontPortFolioEntity getPortFolioInfo = queryFactory
+				.selectFrom(frontPortFolioEntity)
+				.leftJoin(frontPortFolioEntity.commonImageEntityList, commonImageEntity)
+				.fetchJoin()
+				.where(frontPortFolioEntity.idx.eq(existFrontPortFolioEntity.getIdx())
+						.and(frontPortFolioEntity.visible.eq("Y"))
+						.and(commonImageEntity.typeName.eq("portfolio")))
+				.fetchOne();
 
-			return PortFolioMapper.INSTANCE.toDto(getPortFolioInfo);
-		} catch (Exception e) {
-			throw new TspException(NOT_FOUND_PORTFOLIO, e);
-		}
+		return INSTANCE.toDto(getPortFolioInfo);
 	}
 }
