+++ src/main/java/com/tsp/new_tsp_front/api/support/service/FrontSupportJpaApiService.java
@@ -3,8 +3,11 @@ package com.tsp.new_tsp_front.api.support.service;
 import com.tsp.new_tsp_front.api.support.domain.FrontSupportDTO;
 import com.tsp.new_tsp_front.api.support.domain.FrontSupportEntity;
 import com.tsp.new_tsp_front.api.support.service.impl.FrontSupportJpaRepository;
+import com.tsp.new_tsp_front.exception.TspException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+
+import static com.tsp.new_tsp_front.exception.ApiExceptionType.ERROR_SUPPORT;
 
 @Service
 @RequiredArgsConstructor
@@ -21,7 +24,11 @@ public class FrontSupportJpaApiService {
 	 * </pre>
 	 *
 	 */
-	public FrontSupportDTO insertSupportModel(FrontSupportEntity frontSupportEntity) {
-		return this.frontSupportJpaRepository.insertSupportModel(frontSupportEntity);
+	public FrontSupportDTO insertSupportModel(FrontSupportEntity frontSupportEntity) throws TspException {
+		try {
+			return this.frontSupportJpaRepository.insertSupportModel(frontSupportEntity);
+		} catch (Exception e) {
+			throw new TspException(ERROR_SUPPORT, e);
+		}
 	}
 }
